Enable annotation documentation locations (#71)
This PR is motivated by the need to reduce cyclic imports / directly
pass a resolver into the `ResolverKey`.

This PR introduces the `location` argument so you can explicitly say
what the import / documentation path is for a given resolver. This
allows a pre-instantiated resolver to be passed as the second argument
to `ResolverKey` and in some places, avoid cyclic imports in
documentation build due to its dynamic import functionality.